+++ rest/src/main/java/se331/proj/rest/repository/AdvisorRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import se331.proj.rest.entity.Advisor;
 
-public interface AdvisorRepository extends JpaRepository<Advisor,Integer> {
+public interface AdvisorRepository extends JpaRepository<Advisor,Long> {
     List<Advisor> findAll();
 }
